+++ pyfreebilling/settings.py
@@ -199,6 +199,7 @@ INSTALLED_APPS = (
     'bootstrap_pagination',
     'mathfilters',
     'dbbackup',
+    'easy_thumbnails',
 #    'raven.contrib.django.raven_compat',
     #'massadmin',
 )
@@ -321,14 +322,6 @@ SESSION_COOKIE_HTTPONLY = True
 PFB_NB_ADMIN_CDR = 3
 PFB_NB_CUST_CDR = 30
 
-#-----#
-
-try:
-    from .local_settings import *
-except ImportError:
-    pass
-
-#-----#
 
 #-----
 #-- Upload settings
@@ -355,3 +348,94 @@ DBBACKUP_GPG_RECIPIENT = None
 #DBBACKUP_FTP_USER = 'user, blank if anonymous'
 #DBBACKUP_FTP_PASSWORD = 'password, can be blank'
 #DBBACKUP_FTP_PATH = 'path, blank for default'
+
+#Logo settings
+THUMBNAIL_ALIASES = {
+    '': {
+        'avatar': {'size': (0, 50), 'crop': False},
+    },
+}
+
+#########################
+# Customer panel specific
+#########################
+
+# Menu settings #
+# bootstrap add navbar-inverse value to navbar - default ''
+PYFB_CP_MENU_TYPE_INV = '' #navbar-inverse'
+
+# If you have put logo, enter the filename - default ''
+# logo have to be upload in static directory
+# logo 50 px height
+PYFB_CP_LOGO_NAME = ''
+# Company name ALT value
+PYFB_CP_COMPANY_NAME = 'PyFreeBilling'
+
+# Bootstrap and jquery specific files
+BOOTSTRAP3 = {
+
+    # The URL to the jQuery JavaScript file
+    'jquery_url': '//code.jquery.com/jquery.min.js',
+
+    # The Bootstrap base URL
+    'base_url': '//netdna.bootstrapcdn.com/bootstrap/3.3.1/',
+
+    # The complete URL to the Bootstrap CSS file (None means derive it from base_url)
+    #'css_url': None,
+    'css_url': '//maxcdn.bootstrapcdn.com/bootswatch/3.3.0/cosmo/bootstrap.min.css',
+
+    # The complete URL to the Bootstrap CSS file (None means no theme)
+    'theme_url': '//maxcdn.bootstrapcdn.com/bootswatch/3.3.0/cosmo/bootstrap.min.css',
+
+    # The complete URL to the Bootstrap JavaScript file (None means derive it from base_url)
+    'javascript_url': None,
+
+    # Put JavaScript in the HEAD section of the HTML document (only relevant if you use bootstrap3.html)
+    'javascript_in_head': False,
+
+    # Include jQuery with Bootstrap JavaScript (affects django-bootstrap3 template tags)
+    'include_jquery': False,
+
+    # Label class to use in horizontal forms
+    'horizontal_label_class': 'col-md-2',
+
+    # Field class to use in horizontal forms
+    'horizontal_field_class': 'col-md-4',
+
+    # Set HTML required attribute on required fields
+    'set_required': True,
+
+    # Set placeholder attributes to label if no placeholder is provided
+    'set_placeholder': True,
+
+    # Class to indicate required (better to set this in your Django form)
+    'required_css_class': '',
+
+    # Class to indicate error (better to set this in your Django form)
+    'error_css_class': 'has-error',
+
+    # Class to indicate success, meaning the field has valid input (better to set this in your Django form)
+    'success_css_class': 'has-success',
+
+    # Renderers (only set these if you have studied the source and understand the inner workings)
+    'formset_renderers':{
+        'default': 'bootstrap3.renderers.FormsetRenderer',
+    },
+    'form_renderers': {
+        'default': 'bootstrap3.renderers.FormRenderer',
+    },
+    'field_renderers': {
+        'default': 'bootstrap3.renderers.FieldRenderer',
+        'inline': 'bootstrap3.renderers.InlineFieldRenderer',
+    },
+}
+
+
+#-----#
+
+try:
+    from .local_settings import *
+except ImportError:
+    pass
+
+#-----#